pyproject.toml
[tool.poetry]
name = "chemformer"
version = "2.2.0"
description = "Smiles-based synthesis prediction."
authors = ["Irwin, Ross <ross.irwin1@astrazeneca.com>"]
maintainers = [
    "Westerlund, Annie <annie.westerlund@astrazeneca.com>",
    "Saigiridharan, Lakshidaa <lakshidaa.saigiridharan@astrazeneca.com>",
    "Peter Hartog <peter.hartog@astrazeneca.com>",
    "Genheden, Samuel <samuel.genheden@astrazeneca.com>",
]
license = "Apache 2.0"

packages = [{ include = "molbart" }]

[tool.poetry.dependencies]
python = "3.10.14"
rdkit = "^2023.9.5"
torch = "^2.2.2"
pytorch-lightning = "^2.2.1"
pysmilesutils = { git = "https://github.com/MolecularAI/pysmilesutils.git" }
fastapi = "^0.110.0"
pandas = "1.4.3"
deepspeed = "^0.14.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}


[tool.poetry.dev-dependencies]
pre-commit = "*"
pytest = "*"
pytest-mock = "*"
pytest-datadir = "*"
pytest-cov = "*"
black = "*"
isort = "*"
flake8 = "*"
mypy = "*"

[tool.poetry.group.config.dependencies]
hydra-core = "^1.3.2"
hydra-colorlog = "^1.2.0"
hydra-optuna-sweeper = "^1.2.0"

[tool.poetry.group.extras.dependencies]
wandb = "^0.16.5"
textbrewer = "^0.2.1.post1"
eco2ai = "^0.3.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
